将position改为int类型。

+++ components/app/src/main/java/org/mallfoundry/app/Menu.java
@@ -62,7 +62,7 @@ public class Menu implements Position {
     private String title;
 
     @Column(name = "position_")
-    private long position;
+    private int position;
 
     @JsonUnwrapped
     @Embedded
